+++ src/WProgram.h
@@ -1,4 +1,4 @@
 #pragma once
 
 // Compatibility shim for libraries that still include WProgram.h.
-#include <Arduino.h>
+#include "Arduino.h"