+++ src/c70_menuRA.hpp
@@ -175,12 +175,12 @@ void printRASubmenu()
     #if USES_ROTARY_ENCODER == 1
     if (RAselect == 4)
     {
-        lcdMenu.printMenu(">GoTo target");
+        lcdMenu.printMenu(String(">") + TR_GOTO + " " + TR_TARGET);
         return;
     }
     if (RAselect == RA_FIELD_EXIT)
     {
-        lcdMenu.printMenu(">Exit");
+        lcdMenu.printMenu(String(">") + TR_EXIT);
         return;
     }
     #endif