+++ src/wiring.h
@@ -1,4 +1,4 @@
 #pragma once
 
 // Compatibility shim for legacy Arduino libraries.
-#include <Arduino.h>
+#include "Arduino.h"